tests

+++ commands/command-base.js
@@ -1,4 +1,5 @@
 const { prefix, realBot, channels, roles } = require("../config.json").discord
+const { MessageActionRow, MessageSelectMenu } = require("discord.js")
 
 const validatePermissions = (permissions) => {
   const validPermissions = [
@@ -146,6 +147,57 @@ module.exports.listen = (client) => {
         interaction.channel
           .send({
             content: `Thank you for accepting our rules <@${interaction.user.id}>! Head on to <#${channels.welcome}> to choose your role in the server.`,
+          })
+          .then((msg) => {
+            setTimeout(() => msg.delete(), 20000)
+          })
+          .catch(console.error)
+      }
+
+      if (interaction.customId.includes("download")) {
+        interaction.message.components.fetch().then((components) => {
+          console.log(components)
+        })
+
+        let row = new MessageActionRow().addComponents(
+          new MessageSelectMenu()
+            .setCustomId("welcome")
+            .setPlaceholder("Choose an option")
+            .setMinValues(1)
+            .setMaxValues(1)
+            .addOptions([
+              {
+                label: "latest",
+                value: "v0",
+                description: "Version from 26-10-2021",
+              },
+              {
+                label: "v1",
+                value: "v1",
+                description: "Version from 20-10-2021",
+              },
+              {
+                label: "v2",
+                value: "v2",
+                description: "",
+              },
+              {
+                label: "v3",
+                value: "v3",
+                description: "",
+              },
+              {
+                label: "v4",
+                value: "v4",
+                description: "",
+              },
+            ])
+        )
+
+        interaction.channel
+          .send({
+            content: `_ _\n<@${interaction.user.id}> here you have your downloads!\n\nChoose the version of **${interaction.customId}** you would like to download:\n_ _`,
+            components: [row],
           })
           .then((msg) => {
             setTimeout(() => msg.delete(), 20000)
